# Conflicts:
#	.idea/workspace.xml
#	app/src/main/java/fit/budle/model/BudleResponse.kt
#	app/src/main/java/fit/budle/model/EstablishmentParse.kt
#	app/src/main/java/fit/budle/model/EstablishmentStructue.kt
#	app/src/main/java/fit/budle/navigation/NavigationComponent.kt
#	app/src/main/java/fit/budle/network/BudleAPIRequests.kt
#	app/src/main/java/fit/budle/repository/BudleRepository.kt
#	app/src/main/java/fit/budle/viewmodel/MainViewModel.kt

+++ app/src/main/java/fit/budle/model/EstablishmentParse.kt
@@ -1,7 +1,11 @@
 package fit.budle.model
 
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.snapshots.SnapshotStateList
+import fit.budle.R
+
 data class Establishment(
-    val id: Long,
+    val id: Int,
     val name: String,
     val description: String?,
     val address: String,
